feat: Allow dict key access to mapping return values
Currently, if a function used for the Projector returns
an incompatible data type, it is immediately cast to JSON
before returning the value to the format function.
This means notation like `{var[key]}` does not work.
This patch allows to delay that casting to when it is actually
required and thus allows `{var[key]}` to work as expected.