fix(returner): Fix vault events for good (hopefully)
The event system in Salt is a bit convoluted.
When an event is sent via `salt.utils.event.Event.fire_master`,
the master event bus will receive the raw load from the minion
event as data instead of the event itself. This is different
when the event originates from the master daemon via
`salt.utils.event.Event.fire_event`.